This is a dirty commit -- I have supplied updates to "edit.html" which include
a) a basic text editor for python
b) a mechanism to submit the code to a server function

also, in "server.go", i am now able to spawn a new interactive python
shell. Unfortunately, I have not yet hooked up a fifo to it, but that's
on my TODO list.

After that, work on sandboxing the python session? Improving the
front-end? Adding the ability to create new sessions? Etc?